refactor(main-app): adjustment of file name specification
