+++ desktop/main-app/src/bootup/init-env.ts
