+++ desktop/main-app/src/bootup/init-menus.ts
