+++ desktop/main-app/src/utils/bootup-flow.ts
