+++ desktop/main-app/src/utils/close-app.ts
